Updated the extract_data function. here are the changes:
Extracts data from the results of an API call for one particular subreddit at a time.

This is a recursive function that will skip over all mod stickied posts by recursively incrementing the post_count
and start parameters. Post count indicates the number of posts to grab and start indicates which post to start
processing from.